requirements.txt
arrow==0.15.8
asgiref==3.2.3
Babel==2.8.0
baidu-aip==2.2.18.0
binaryornot==0.4.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cryptography==2.9.2
Django==3.0.3
django-cors-headers==3.4.0
django-filter # @ file:///home/conda/feedstock_root/build_artifacts/django-filter_1591380509061/work
django-ipware==3.0.0
django-jazzmin==2.2.5
django-phonenumber-field==4.0.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.10
importlib-metadata # @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1593211382425/work
Jinja2==2.11.2
jinja2-time==0.2.0
Markdown # @ file:///home/conda/feedstock_root/build_artifacts/markdown_1589366472132/work
markdown2==2.3.9
MarkupSafe==1.1.1
olefile==0.46
phonenumbers==8.12.7
Pillow # @ file:///Users/runner/miniforge3/conda-bld/pillow_1593693221521/work
poyo==0.5.0
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
zipp==3.1.0
